+++ CHANGELOG.md
@@ -10,6 +10,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 For release notes before version 3.1, please refer to https://github.com/Accenture/mercury
 
 ---
+## Version 4.0.8, 10/9/2024
+
+### Added
+
+1. Partial support of Active Profile using the "spring.profiles.active" parameter
+2. Hierarchy of flows
+
+### Removed
+
+N/A
+
+### Changed
+
+N/A
+
+---
 ## Version 4.0.7, 10/1/2024
 
 ### Added